requirements.txt
pydantic>=2.0.0
Pillow>=9.0.0
python-dotenv>=0.19.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
aiofiles>=23.0.0
requests>=2.31.0